+++ petshelter/src/main/java/com/ironhack/petshelter/repository/AdopterRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AdopterRepository extends JpaRepository<Adopter, Integer> {
+public interface AdopterRepository extends JpaRepository<Adopter, Long> {
 }
